requirements.txt
Django==5.1.3
black==24.10.0
isort==5.13.2
pylint==3.3.1
crispy-tailwind==1.0.3
celery==5.4.0
redis==5.2.0
djangorestframework==3.15.2
django-extensions==3.2.3
psycopg2-binary==2.9.10
django-environ==0.11.2
drf-spectacular==0.27.2